:hammer: Fixed index conversion bug
See https://github.com/Gjacquenot/test_clang/runs/7747996225?check_suite_focus=true

[615/908] Building CXX object grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o
FAILED: grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o
/usr/bin/clang++ -DGTEST_HAS_POSIX_RE=0 -I. -I/opt/share/code/grpc/inc -I/opt/share/code/ssc -I/opt/share/code/exceptions/inc -I/opt/share/code/external_data_structures/inc -I/opt/share/code/core/inc -I/opt/share/code/environment_models/inc -I/opt/share/code/hdb_interpolators/inc -I/opt/share/code/mesh/inc -isystem /opt/eigen -isystem /opt/boost/include -isystem /opt/grpc/include -isystem grpc -isystem /opt/share/code/yaml-cpp/src -isystem /opt/share/code/yaml-cpp/include -isystem /opt/share/code/yaml-cpp/include/yaml-cpp -Werror -pedantic -pedantic-errors -Wextra -Wall -Wunused-function -Wunused-label -Wunused-parameter -Wunused-value -Wunused-variable -fno-common -Wformat=2 -Winit-self -Wpacked -Wpointer-arith -Wmissing-declarations -Wmissing-format-attribute -Wsign-compare -Wstrict-aliasing=2 -Wundef -Wconversion -Wno-misleading-indentation -Woverloaded-virtual -Weffc++ -Wwrite-strings -Wfatal-errors -Wno-deprecated -Wvariadic-macros -fPIC -std=gnu++11 -O3 -DNDEBUG -MD -MT grpc/CMakeFiles/grpc.dir/src/Surf...
/opt/share/code/grpc/src/SurfaceElevationFromGRPC.cpp:383:36: fatal error: implicit conversion changes signedness: 'int' to 'std::vector::size_type' (aka 'unsigned long') [-Wsign-conversion]
                        s.phase.at(j).reserve(static_cast<size_t>(response.spectrum(i).phase(j).phase_size()));
                                ~~ ^
1 error generated.
[616/908] Building CXX object force_models/unit_tests/CMakeFiles/force_models_tests.dir/src/maneuvering_parserTest.cpp.o
[617/908] Building CXX object grpc/CMakeFiles/grpc.dir/src/GRPCForceModel.cpp.o
ninja: build stopped: subcommand failed.
make: *** [Makefile:62: build-debian-clang] Error 1
Error: Process completed with exit code 2.

+++ code/grpc/src/SurfaceElevationFromGRPC.cpp
@@ -380,8 +380,8 @@ class SurfaceElevationFromGRPC::Impl
                     s.phase.resize(static_cast<size_t>(response.spectrum(i).phase_size()));
                     for (int j = 0 ; j < response.spectrum(i).phase_size() ; ++j)
                     {
-                        s.phase.at(j).reserve(static_cast<size_t>(response.spectrum(i).phase(j).phase_size()));
-                        std::copy(response.spectrum(i).phase(j).phase().begin(), response.spectrum(i).phase(j).phase().end(), std::back_inserter(s.phase.at(j)));
+                        s.phase.at(static_cast<size_t>(j)).reserve(static_cast<size_t>(response.spectrum(i).phase(j).phase_size()));
+                        std::copy(response.spectrum(i).phase(j).phase().begin(), response.spectrum(i).phase(j).phase().end(), std::back_inserter(s.phase.at(static_cast<size_t>(j))));
                     }
                 }
                 ret.push_back(s);
